+++ backend/app/api/v1/endpoints/posts.py
@@ -1,6 +1,7 @@
 from typing import Any, List
-
 from fastapi import APIRouter, HTTPException
+from crud import crud_posts
+
 
 router = APIRouter()
 
@@ -8,7 +9,7 @@ router = APIRouter()
 @router.get("/")
 def get_posts():
     try:
-        return {}
+        return crud_posts.CRUD_posts.get_all
     except HTTPException:
         pass 
 
@@ -16,7 +17,7 @@ def get_posts():
 @router.get("/{id}")
 def get_post_by_id(id):
     try:
-        return {}
+        return crud_posts.CRUD_posts.get_by_id(id)
     except HTTPException:
         pass 
 
@@ -24,7 +25,7 @@ def get_post_by_id(id):
 @router.post("/")
 def create_post():
     try:
-        return {}
+        return crud_posts.CRUD_posts.create
     except HTTPException:
         pass 
 
@@ -32,7 +33,7 @@ def create_post():
 @router.patch("/{id}")
 def modify_post():
     try:
-        return {}
+        return crud_posts.CRUD_posts.modify(id)
     except HTTPException:
         pass 
 
@@ -40,6 +41,6 @@ def modify_post():
 @router.delete("/{id}")
 def delete_post(id):
     try:
-        return {}
+        return crud_posts.CRUD_posts.delete(id)
     except HTTPException:
         pass